+++ GoToDnSpy/ReleaseNotes.md
@@ -1,4 +1,11 @@
 # Release notes
 
+- 1.0.2
+    + Added navigate support for Roslyn `PropertyDeclarationSyntax`  
+        (less "is not a valid identifier" errors)
+    + Added support for solution projects output, now we can easy see CIL of compiled project! 
+        (must build project first)
+- 1.0.1
+    * Fix error if empty dnSpy path
 - 1.0
-    First public release
+    + First public release